pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kodex"
version = "3.0.0"
description = "A blazing-fast text expansion / hotstring engine for Windows (and eventually cross-platform)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "SilverSix311" },
]
keywords = ["text-expansion", "hotstring", "productivity", "typing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing",
]

dependencies = [
    "pynput>=1.7.6",
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
    "pyperclip>=1.8.2",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
kodex = "kodex_py.cli:cli"
kodex-migrate = "kodex_py.storage.migration:main"

[tool.hatch.build.targets.wheel]
packages = ["src/kodex_py"]

[tool.hatch.build.targets.sdist]
include = ["src/kodex_py", "resources"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
